analyzer/requirements.txt
pytesseract
pdf2image
Pillow
requests
python-dotenv
numpy
streamlit

# Chandra OCR — base package (vLLM/OpenAI-compatible server). For local HF weights: pip install 'chandra-ocr[hf]'
chandra-ocr>=0.2.0
